Solve circular imports while typing
