+++ falcon/http_error.py
@@ -21,10 +21,10 @@ from typing import Union
 import xml.etree.ElementTree as et
 
 from falcon.constants import MEDIA_JSON
-from falcon.link import Link
-from falcon.media_serializer import Serializer
-from falcon.typing_http_data import RawHeaders
-from falcon.typing_http_data import Status
+from falcon.typing import Link
+from falcon.typing import Serializer
+from falcon.typing import RawHeaders
+from falcon.typing import Status
 from falcon.util import code_to_http_status, http_status_to_code, uri
 from falcon.util.deprecation import deprecated_args
