--- falcon/link.py
@@ -1,3 +1,0 @@
-from typing import Dict
-
-Link = Dict[str, str]
